+++ src/commands/download/downloadSet.ts
@@ -1,0 +1,25 @@
+import { AxiosInstance } from "axios";
+import colors from "colors";
+import yargs = require("yargs");
+import { getSetCDN, getSetJSON, saveSet } from "../../functions";
+import SaveSetOptions from "../../SaveSetOptions";
+
+export default async function downloadSet(setId: number, api: AxiosInstance, argv: yargs.Arguments) {
+  const outputPath = argv.output;
+  const options = new SaveSetOptions({ downloadImages: argv.p, redownloadImages: argv.r, fragmentCards: argv.f, log: argv.l });
+  // Make request for sets
+  return await getSetCDN(api, setId)
+    .then((cdn) => {
+      return getSetJSON(cdn)
+        .then((set) => saveSet(set, outputPath, options))
+        .then((value) => {
+          console.log(colors.blue(setId + ":"), `Success! Set #${setId} downloaded.`);
+          return value;
+        })
+        .catch((err) => {
+          console.log("Error getting set " + cdn.setId, err);
+          return false;
+        });
+    })
+    .catch((err) => false);
+}
